% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appg_functions.R
\name{fetch_appg_officers}
\alias{fetch_appg_officers}
\title{\code{fetch_appg_officers} Downloads data on APPG officers and generates new name variable}
\usage{
fetch_appg_officers(date, mp_noble_names = c())
}
\arguments{
\item{date}{The date of the register to download ("yyyy-mm-dd")}

\item{mp_noble_names}{A vector of names whose beginning equals a noble title, e.g. "Earl Jones"}
}
\value{
A tibble incuding data on APPGs, their officers and names for joining with other data
}
\description{
\code{fetch_appg_officers} Downloads data on APPG officers and generates new name variable
}
