% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edms_functions.R
\name{fetch_edm_data}
\alias{fetch_edm_data}
\title{\code{fetch_edm_data} Download and process data on EDMs including signatures}
\usage{
fetch_edm_data(start_date, end_date)
}
\arguments{
\item{start_date}{The start date of the period of interest ("yyyy-mm-dd")}

\item{end_date}{The end date of the period of interest ("yyyy-mm-dd")}
}
\value{
A tibble containing data on signatures and the primary sponsor of an EDM
}
\description{
\code{fetch_edm_data} Download and process data on EDMs including signatures
}
