% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appg_functions.R
\name{get_appg_full_name}
\alias{get_appg_full_name}
\title{\code{get_appg_full_name} Creates new name variable to enable joining with Member data}
\usage{
get_appg_full_name(officers)
}
\arguments{
\item{officers}{A tibble with data on APPGs and their officers}
}
\value{
A tibble incuding the above plus a new name variable
}
\description{
\code{get_appg_full_name} Creates new name variable to enable joining with Member data
}
\keyword{internal}
