% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard_debates_functions.R
\name{get_debates}
\alias{get_debates}
\title{\code{get_debates} Get the number of debates each Member spoke in during a defined period}
\usage{
get_debates(members = MEMBERS, start_date, end_date = Sys.Date())
}
\arguments{
\item{members}{A tibble with Member data, including a MNIS ID (constant given as default)}

\item{start_date}{The start date of the period of interest ("yyyy-mm-dd")}

\item{end_date}{The end date of the period of interest ("yyyy-mm-dd"), set to today as default}
}
\value{
A tibble incuding for each current MP their url and the total number of debates they spoke in
}
\description{
\code{get_debates} Get the number of debates each Member spoke in during a defined period
}
