% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_functions.R
\name{get_div_ids}
\alias{get_div_ids}
\title{Find all divisions and their IDs within a date range}
\usage{
get_div_ids(start_date, end_date)
}
\arguments{
\item{start_date}{The start date of the period of interest ("yyyy-mm-dd")}

\item{end_date}{The end date of the period of interest ("yyyy-mm-dd")}
}
\value{
A vector containing division IDs
}
\description{
\code{get_div_ids} Gets division IDs for all divisions in date range
}
\keyword{internal}
