% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_functions.R
\name{get_division}
\alias{get_division}
\title{Download data by division}
\usage{
get_division(division_id)
}
\arguments{
\item{division_id}{The unique ID of the division to download}
}
\value{
A tibble containing data on a division
}
\description{
\code{get_division} Downloads raw data on division
}
\keyword{internal}
