% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_functions.R
\name{get_divisions_data}
\alias{get_divisions_data}
\title{Download and process division data}
\usage{
get_divisions_data(start_date, end_date)
}
\arguments{
\item{start_date}{The start date of the period of interest ("yyyy-mm-dd")}

\item{end_date}{The end date of the period of interest ("yyyy-mm-dd")}
}
\value{
A tibble showing for all MPs voting Aye and No, all MPs who did not vote and all Tellers for all divisions within a date range
}
\description{
\code{get_divisions_data} Downloads and processes division data to show votes by MP
}
