% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edms_functions.R
\name{get_edms_list}
\alias{get_edms_list}
\title{Fetch and manipulate data on EDMs signed by Members}
\usage{
get_edms_list(start_date, end_date)
}
\arguments{
\item{start_date}{The start date of the period of interest ("yyyy-mm-dd")}

\item{end_date}{The end date of the period of interest ("yyyy-mm-dd")}
}
\value{
A tibble containing data on EDMs and their IDs
}
\description{
\code{get_edms_list} Downloads a list of all EDMs signed in a period comprising whole weeks
}
\keyword{internal}
