% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edms_functions.R
\name{get_edms_signatures}
\alias{get_edms_signatures}
\title{\code{get_edms_signatures} Gets all signatures for a given EDM}
\usage{
get_edms_signatures(id)
}
\arguments{
\item{id}{The unique identifier for an EDM}
}
\value{
A tibble containing data on signatures and the primary sponsor of an EDM
}
\description{
\code{get_edms_signatures} Gets all signatures for a given EDM
}
\keyword{internal}
