% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_full_name.R
\name{get_full_name}
\alias{get_full_name}
\title{\code{get_full_name} Creates new name variable to enable joining Parliamentary search data with Member data}
\usage{
get_full_name(names)
}
\arguments{
\item{names}{A vector with Members' names (in activity data from Parliamentary Search)}
}
\value{
A 'full name' vector with Members' names in standardised format
}
\description{
\code{get_full_name} Creates new name variable to enable joining Parliamentary search data with Member data
}
\keyword{export}
