% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles_functions.R
\name{get_gvt_posts}
\alias{get_gvt_posts}
\title{Fetch data on the government and opposition posts MPs hold}
\usage{
get_gvt_posts(members = MEMBERS)
}
\arguments{
\item{members}{A tibble with Member data, including a MNIS ID (constant given as default)}
}
\value{
A tibble showing the government roles MPs hold
}
\description{
\code{get_gvt_posts} Downloads data on government posts
}
\keyword{Internal}
