% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_joining_functions.R
\name{get_match_names}
\alias{get_match_names}
\title{Functions to join parliamentary search data with Member data using fuzzy matching}
\usage{
get_match_names(members = MEMBERS)
}
\arguments{
\item{members}{A tibble with data on current Members (constant provided as default)}

\item{missing_name}{A vector of Member names in activity data that do not match names in Member data}
}
\value{
A function to use fuzzy matching to find matches for Member names
}
\description{
\code{get_match_names} Creates a function to use fuzzy matching to find matches for Member names
}
\keyword{internal}
