% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_members.R
\name{get_member_data}
\alias{get_member_data}
\title{Functions to download data on Members from MNIS}
\usage{
get_member_data()
}
\value{
A tibble with data on current Members
}
\description{
\code{get_member_data} Downloads Member data from MNIS
}
\keyword{internal}
