% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_members.R
\name{get_member_full_name}
\alias{get_member_full_name}
\title{\code{get_member_full_name} Standardises Members' names for joining with other data}
\usage{
get_member_full_name(members)
}
\arguments{
\item{members}{A tibble with data on current Members}
}
\value{
A tibble with data on current Members, including a vector with standardised names
}
\description{
\code{get_member_full_name} Standardises Members' names for joining with other data
}
\keyword{internal}
