% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_members.R
\name{get_members}
\alias{get_members}
\title{\code{get_members} Downloads and manipulates data on current Members}
\usage{
get_members()
}
\value{
A tibble with data on current Members, including a vector with standardised names
}
\description{
\code{get_members} Downloads and manipulates data on current Members
}
