% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard_debates_functions.R
\name{get_number}
\alias{get_number}
\title{\code{get_number} Find the number of debates an MP spoke in}
\usage{
get_number(url, pb)
}
\arguments{
\item{url}{The url showing the number of debates}

\item{pb}{Generates a progress bar using the length of the vector of urls}
}
\value{
A tibble incuding an MP's url and the total number of debates they spoke in
}
\description{
\code{get_number} Find the number of debates an MP spoke in
}
\keyword{internal}
