% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appg_functions.R
\name{get_officer_data}
\alias{get_officer_data}
\title{Fetch and manipulate data on APPG officers}
\usage{
get_officer_data(date, mp_noble_names = c())
}
\arguments{
\item{date}{The date of the register to download ("yyyy-mm-dd")}

\item{mp_noble_names}{A vector of names whose beginning equals a noble title, e.g. "Earl Jones"}
}
\value{
A tibble containing data on APPGs and their officers
}
\description{
\code{get_officer_data} Downloads data on MPs holding roles in APPGs
}
\keyword{internal}
