% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appg_functions.R
\name{get_officers}
\alias{get_officers}
\title{Download raw data on APPG officers}
\usage{
get_officers(register_date)
}
\arguments{
\item{date}{The date of the register to download ("yyyy-mm-dd")}

\item{mp_noble_names}{A vector of names whose beginning equals a noble title, e.g. "Earl Jones"}
}
\value{
A tibble containing data on APPGs and their officers
}
\description{
\code{get_officer_data} Downloads raw data on APPG officers
}
\keyword{internal}
