% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles_functions.R
\name{get_opp_posts}
\alias{get_opp_posts}
\title{\code{get_opp_posts} Downloads data on opposition posts}
\usage{
get_opp_posts(members = MEMBERS)
}
\arguments{
\item{members}{A tibble with Member data, including a MNIS ID (constant given as default)}
}
\value{
A tibble showing the opposition roles MPs hold
}
\description{
\code{get_opp_posts} Downloads data on opposition posts
}
\keyword{Internal}
