% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top100.R
\name{get_top100}
\alias{get_top100}
\title{\code{get_top100} Gets the 100 most spoken words for each MP}
\usage{
get_top100(contributions_joined, members = MEMBERS)
}
\arguments{
\item{contributions_joined}{A tibble with data on Members' contributions (from parliamentary search), joined with Member data using this package's joining functions}

\item{members}{A tibble with Member data (constant given as default)}
}
\value{
A tibble with each Member's 100 most used words
}
\description{
\code{get_top100} Gets the 100 most spoken words for each MP
}
