% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard_debates_functions.R
\name{get_url}
\alias{get_url}
\title{Fetch data on how many debates MPs took part in}
\usage{
get_url(mnis_id, start_date, end_date = Sys.Date())
}
\arguments{
\item{mnis_id}{An MP's MNIS identification number}

\item{start_date}{The start date of the period of interest ("yyyy-mm-dd")}

\item{end_date}{The end date of the period of interest ("yyyy-mm-dd"), set to today as default}
}
\value{
A tibble incuding an MP's MNIS number and their url
}
\description{
\code{get_url} Generates url showing debates MP spoke in during a defined period
}
\keyword{internal}
