% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_joining_functions.R
\name{join_and_match}
\alias{join_and_match}
\title{\code{join_and_match} Joins parliamentary search data with Member data using fuzzy matching}
\usage{
join_and_match(data, members = MEMBERS)
}
\arguments{
\item{data}{A tibble with activity data from Parliamentary Search (may also work on other data)}

\item{members}{A tibble with data on current Members (constant provided as default)}
}
\value{
A list including a tibble with activity data joined with Member data (joined_data) and a tibble showing names matched using fuzzy matching (matches)
}
\description{
\code{join_and_match} Joins parliamentary search data with Member data using fuzzy matching
}
