% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions_functions.R
\name{process_division}
\alias{process_division}
\title{Process division data}
\usage{
process_division(division)
}
\arguments{
\item{division}{Raw data for the division of interest}
}
\value{
A tibble showing for a division all MPs voting Aye and No, all MPs who did not vote and all Tellers
}
\description{
\code{process_division} Processes raw division data to show votes by MP
}
\keyword{internal}
