% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles_functions.R
\name{get_posts}
\alias{get_posts}
\title{\code{get_posts} Downloads data on government and opposition posts}
\usage{
get_posts(members = MEMBERS)
}
\arguments{
\item{members}{A tibble with Member data, including a MNIS ID (constant given as default)}
}
\value{
A tibble showing the government and opposition roles MPs hold
}
\description{
\code{get_posts} Downloads data on government and opposition posts
}
